requirements.txt
torch>=2.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
psycopg2-binary>=2.9
streamlit>=1.28
plotly>=5.17
python-dotenv>=1.0
sqlalchemy>=2.0
joblib>=1.3
